feat(web): abstract-claim-pnk-button-add-in-get-pnk-page
